requirements.txt
tensorflow-gpu==1.14
keras==2.2.5
matplotlib==3.1.1
pandas==0.25.1
numpy==1.17.1
